added tests for SpriteButtonComponent and removed custom_sprite_test file
